Move/merge some deterministic ID tests from the observed-data SDO test suite to the test suite specific to testing deterministic IDs.  This keeps the tests for that specific system in one place.
